test team authentication is deleted in case token was revoked

+++ app/models/team/find_invitable_slack_users.rb
@@ -18,9 +18,7 @@ class Team::FindInvitableSlackUsers
       begin
         Slack::Web::Client.new(token: team_authentication&.token).users_list.members
       rescue Slack::Web::Api::Error => exception
-        if exception.message == "token_revoked"
-          team_authentication.destroy
-        end
+        team_authentication.destroy if exception.message == "token_revoked"
 
         Rails.logger.error("Team::FindInvitableSlackUsers#all_slack_members failed (team.id=#{@team.id}) with (#{exception.class}=#{exception.message})")
         Raven.capture_exception(exception, extra: { team_id: @team.id })
